+++ lib/capybara/rspec/matchers/have_selector.rb
@@ -1,11 +1,14 @@
 # frozen_string_literal: true
 
 require 'capybara/rspec/matchers/base'
+require 'capybara/rspec/matchers/count_sugar'
 
 module Capybara
   module RSpecMatchers
     module Matchers
       class HaveSelector < WrappedElementMatcher
+        include CountSugar
+
         def element_matches?(el)
           el.assert_selector(*@args, &@filter_block)
         end